- move implementation out of client
